backend/requirements.txt
# HR Recruitment Agent — backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
greenlet>=3.1.0
httpx>=0.27.0
langgraph>=0.2.50
langchain-core>=0.3.20
langchain-openai>=0.2.8
numpy>=2.0.0
python-jose[cryptography]>=3.3.0
aiosmtplib>=3.0.0
email-validator>=2.2.0
pypdf>=5.0.0
python-docx>=1.1.0
